backend/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
python-dotenv==1.0.1
google-generativeai==0.8.3
qdrant-client==1.12.0
psycopg2-binary==2.9.10
pydantic==2.10.0
PyJWT==2.9.0
